Add 'position' field to Vet and enable search by position
This commit adds a 'position' field to the Vet model in order to enable search of vets by their professional position.
Consequently, an endpoint '/position/{position}' was created in VetResource, which retrieves a list of vet objects whose position matches the input string.
The VetRepository is updated with a new method findVetByPosition, and a corresponding test is added to VetControllerTest.
The schema update script file 'change.sql' is added under resources to upgrade the database structure. A minor improvement is also made in the README file for better command formatting.